pyproject.toml
[tool.poetry]
name = "fast-soy-admin"
version = "0.1.0"
description = ""
authors = ["kk <kanbang@163.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
uvicorn = "^0.29.0"
fastapi = "^0.111.0"
pydantic = "^2.7.1"
tortoise-orm = "^0.20.1"
pydantic-settings = "^2.2.1"
loguru = "^0.7.2"
pyjwt = "^2.8.0"
passlib = "^1.7.4"
aerich = "^0.7.2"
argon2-cffi = "^23.1.0"
apscheduler = "^3.10.4"
sqlalchemy = "^2.0.31"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
